move interception points back to limit
